@@ -4,26 +4,27 @@
   "private": true,
   "type": "module",
   "devDependencies": {
-    "@babel/cli": "7.23.0",
-    "@babel/core": "7.23.3",
+    "@babel/cli": "7.23.4",
+    "@babel/core": "7.23.6",
     "@babel/plugin-proposal-class-properties": "7.18.6",
     "@babel/plugin-proposal-pipeline-operator": "7.23.3",
-    "@babel/preset-env": "7.23.3",
+    "@babel/preset-env": "7.23.6",
     "@babel/preset-flow": "7.23.3",
-    "@cucumber/cucumber": "10.0.1",
+    "@cucumber/cucumber": "10.1.0",
     "@sveltejs/adapter-auto": "2.1.1",
-    "@sveltejs/kit": "1.27.5",
-    "@sveltejs/vite-plugin-svelte": "2.5.2",
+    "@sveltejs/kit": "1.30.3",
+    "@sveltejs/vite-plugin-svelte": "2.5.3",
     "@testing-library/dom": "9.3.3",
-    "@testing-library/jest-dom": "6.1.4",
+    "@testing-library/jest-dom": "6.1.6",
     "@testing-library/svelte": "4.0.5",
     "@tsconfig/svelte": "5.0.2",
     "@types/chai-jest-snapshot": "1.3.8",
-    "@types/node": "20.9.0",
-    "@types/semver": "7.5.5",
+    "@types/node": "20.10.5",
+    "@types/semver": "7.5.6",
     "@types/tap": "15.0.11",
+    "@vitest/browser": "1.1.0",
     "ajv": "8.12.0",
-    "axios": "1.6.1",
+    "axios": "1.6.3",
     "chai": "4.3.10",
     "chai-as-promised": "7.1.1",
     "chai-jest-snapshot": "2.0.0",
@@ -36,7 +37,7 @@
     "jest": "29.7.0",
     "jest-environment-jsdom": "29.7.0",
     "jest-environment-jsdom-sixteen": "1.0.3",
-    "jsdom": "22.1.0",
+    "jsdom": "23.0.1",
     "karma": "6.4.2",
     "karma-chai": "0.1.0",
     "karma-chrome-launcher": "3.2.0",
@@ -46,22 +47,22 @@
     "karma-webpack": "5.0.0",
     "minimatch": "9.0.3",
     "mocha": "10.2.0",
-    "mutation-testing-metrics": "2.0.5",
-    "mutation-testing-report-schema": "2.0.5",
+    "mutation-testing-metrics": "3.0.0",
+    "mutation-testing-report-schema": "3.0.0",
+    "playwright": "1.40.1",
     "rxjs": "7.8.1",
     "semver": "7.5.4",
-    "svelte": "4.2.3",
-    "svelte-check": "3.6.0",
+    "svelte": "4.2.8",
+    "svelte-check": "3.6.2",
     "svelte-jester": "3.0.0",
-    "tap": "18.5.8",
+    "tap": "18.6.1",
     "ts-jest": "29.1.1",
-    "ts-loader": "9.5.0",
-    "ts-node": "10.9.1",
+    "ts-loader": "9.5.1",
+    "ts-node": "10.9.2",
     "tslib": "2.6.2",
-    "typescript": "5.2.2",
-    "vite": "4.5.0",
-    "vitest": "0.34.6",
-    "vitest-dom": "0.1.1",
+    "typescript": "5.3.3",
+    "vite": "4.5.1",
+    "vitest": "1.1.0",
     "webpack": "5.89.0",
     "webpack-cli": "5.1.4"
   },
