@@ -8,46 +8,47 @@
   "devDependencies": {
     "@types/chai-as-promised": "7.1.8",
     "@types/chai-jest-snapshot": "1.3.8",
-    "@types/istanbul": "0.4.33",
-    "@types/karma": "6.3.7",
-    "@types/mocha": "10.0.4",
-    "@types/node": "20.9.0",
-    "@types/sinon": "17.0.1",
+    "@types/istanbul": "0.4.34",
+    "@types/karma": "6.3.8",
+    "@types/mocha": "10.0.6",
+    "@types/node": "20.10.5",
+    "@types/sinon": "17.0.2",
     "@types/sinon-chai": "3.2.12",
     "@types/source-map-support": "0.5.10",
-    "@typescript-eslint/eslint-plugin": "6.10.0",
-    "@typescript-eslint/parser": "6.10.0",
+    "@typescript-eslint/eslint-plugin": "6.14.0",
+    "@typescript-eslint/parser": "6.14.0",
     "c8": "8.0.1",
     "chai": "4.3.10",
     "chai-as-promised": "7.1.1",
     "chai-jest-snapshot": "2.0.0",
     "concurrently": "8.2.2",
     "cross-env": "7.0.3",
-    "eslint": "8.53.0",
-    "eslint-config-prettier": "9.0.0",
+    "eslint": "8.56.0",
+    "eslint-config-prettier": "9.1.0",
     "eslint-import-resolver-typescript": "3.6.1",
-    "eslint-plugin-import": "2.29.0",
-    "eslint-plugin-prettier": "5.0.1",
-    "eslint-plugin-unicorn": "49.0.0",
+    "eslint-plugin-import": "2.29.1",
+    "eslint-plugin-prettier": "5.1.2",
+    "eslint-plugin-unicorn": "50.0.1",
     "execa": "8.0.1",
     "glob": "10.3.10",
     "install-local": "3.0.1",
     "jasmine": "5.1.0",
     "jasmine-core": "5.1.1",
     "json-schema-to-typescript": "13.1.1",
-    "lerna": "7.4.2",
+    "lerna": "8.0.1",
     "minimatch": "9.0.3",
     "mocha": "10.2.0",
-    "prettier": "3.0.3",
+    "prettier": "3.1.1",
     "rimraf": "5.0.5",
     "sinon": "17.0.1",
     "sinon-chai": "3.7.0",
     "source-map-support": "0.5.21",
-    "ts-node": "10.9.1",
-    "typescript": "5.2.2"
+    "ts-node": "10.9.2",
+    "typescript": "5.3.3"
   },
   "scripts": {
     "all": "npm run clean && npm run build && npm run lint && npm run test",
+    "postinstall": "playwright install chromium",
     "lint": "eslint . --resolve-plugins-relative-to . --ext .ts,.js",
     "lint:fix": "npm run lint -- --fix",
     "clean": "rimraf --glob \"packages/*/{.nyc_output,reports,coverage,src-generated,*.tsbuildinfo,.stryker-tmp,dist,testResources/tmp}\" \"packages/core/schema/stryker-schema.json\"",
@@ -58,7 +59,7 @@
     "download-incremental-reports": "./tasks/download-incremental-reports.sh",
     "test:mutation:incremental": "lerna run stryker --stream --concurrency 1 -- --incremental",
     "e2e": "npm run e2e:install && npm run e2e:lint && npm run e2e:run",
-    "e2e:install": "cd e2e && npm ci",
+    "e2e:install": "cd e2e && npm ci --force",
     "e2e:lint": "cd e2e && npm run lint",
     "e2e:run": "cd e2e && npm t",
     "e2e:update-snapshots": "CHAI_JEST_SNAPSHOT_UPDATE_ALL=true npm run e2e",
@@ -68,6 +69,6 @@
     "version": "node tasks/bump-peer-deps.js"
   },
   "engines": {
-    "node": ">=16.0.0"
+    "node": ">=18.0.0"
   }
 }
