@@ -1,6 +1,6 @@
 {
   "name": "@stryker-mutator/cucumber-runner",
-  "version": "7.3.0",
+  "version": "8.0.0",
   "description": "A plugin to use the cucumber test runner in Stryker, the JavaScript mutation testing framework",
   "files": [
     "dist/schema",
@@ -29,7 +29,7 @@
     "directory": "packages/cucumber-runner"
   },
   "engines": {
-    "node": ">=16.0.0"
+    "node": ">=18.0.0"
   },
   "keywords": [
     "stryker",
@@ -46,18 +46,18 @@
   "author": "Nico Jansen <jansennico@gmail.com>",
   "license": "ISC",
   "devDependencies": {
-    "@cucumber/cucumber": "10.0.1",
-    "@stryker-mutator/test-helpers": "7.3.0",
-    "@types/semver": "7.5.5"
+    "@cucumber/cucumber": "10.1.0",
+    "@stryker-mutator/test-helpers": "8.0.0",
+    "@types/semver": "7.5.6"
   },
   "peerDependencies": {
     "@cucumber/cucumber": ">=8.0.0",
-    "@stryker-mutator/core": "~7.3.0"
+    "@stryker-mutator/core": "~8.0.0"
   },
   "dependencies": {
-    "@cucumber/messages": "^23.0.0",
-    "@stryker-mutator/api": "7.3.0",
-    "@stryker-mutator/util": "7.3.0",
+    "@cucumber/messages": "^24.0.0",
+    "@stryker-mutator/api": "8.0.0",
+    "@stryker-mutator/util": "8.0.0",
     "semver": "^7.3.7"
   }
 }
