@@ -1,6 +1,6 @@
 {
   "name": "@stryker-mutator/instrumenter",
-  "version": "7.3.0",
+  "version": "8.0.0",
   "description": "The code instrumenter used in Stryker, the JavaScript mutation testing framework",
   "main": "dist/src/index.js",
   "type": "module",
@@ -25,7 +25,7 @@
     "directory": "packages/instrumenter"
   },
   "engines": {
-    "node": ">=16.0.0"
+    "node": ">=18.0.0"
   },
   "publishConfig": {
     "access": "public"
@@ -46,19 +46,19 @@
     "@babel/parser": "~7.23.0",
     "@babel/plugin-proposal-decorators": "~7.23.0",
     "@babel/preset-typescript": "~7.23.0",
-    "@stryker-mutator/api": "7.3.0",
-    "@stryker-mutator/util": "7.3.0",
-    "angular-html-parser": "~5.0.0",
+    "@stryker-mutator/api": "8.0.0",
+    "@stryker-mutator/util": "8.0.0",
+    "angular-html-parser": "~5.2.0",
     "semver": "~7.5.4",
     "weapon-regex": "~1.1.0"
   },
   "devDependencies": {
     "@babel/preset-react": "7.23.3",
-    "@stryker-mutator/test-helpers": "7.3.0",
-    "@types/babel__core": "7.20.4",
-    "@types/babel__generator": "7.6.7",
+    "@stryker-mutator/test-helpers": "8.0.0",
+    "@types/babel__core": "7.20.5",
+    "@types/babel__generator": "7.6.8",
     "@types/estree": "1.0.5",
     "babel-plugin-transform-decorators-legacy": "1.3.5",
-    "svelte": "4.2.3"
+    "svelte": "4.2.8"
   }
 }
