@@ -1,13 +1,11 @@
-{
-  const chai = globalThis.chai ?? require('chai');
+import chai from 'chai';
 
-  globalThis.expect = chai.expect;
-  chai.util.addMethod(chai.Assertion.prototype, 'toEqual', function (expected) {
-    var obj = chai.util.flag(this, 'object');
-    new chai.Assertion(obj).to.deep.equal(expected);
-  });
-  chai.util.addMethod(chai.Assertion.prototype, 'toBe', function (expected) {
-    var obj = chai.util.flag(this, 'object');
-    new chai.Assertion(obj).to.equal(expected);
-  });
-}
+globalThis.expect = chai.expect;
+chai.util.addMethod(chai.Assertion.prototype, 'toEqual', function (expected) {
+  var obj = chai.util.flag(this, 'object');
+  new chai.Assertion(obj).to.deep.equal(expected);
+});
+chai.util.addMethod(chai.Assertion.prototype, 'toBe', function (expected) {
+  var obj = chai.util.flag(this, 'object');
+  new chai.Assertion(obj).to.equal(expected);
+});
