@@ -1,16 +1,13 @@
-{
-  const concat = globalThis.concat ?? require('../src/concat').concat;
-  const greet = globalThis.greet ?? require('../src/concat').greet;
+import { concat, greet } from '../src/concat.js';
 
-  describe('concat', () => {
-    it('should concat a and b', () => {
-      expect(concat('foo', 'bar')).toBe('foobar');
-    });
+describe('concat', () => {
+  it('should concat a and b', () => {
+    expect(concat('foo', 'bar')).toBe('foobar');
   });
+});
 
-  describe('greet', () => {
-    it('should greet me', () => {
-      expect(greet('me')).toBe('👋 me')
-    });
+describe('greet', () => {
+  it('should greet me', () => {
+    expect(greet('me')).toBe('👋 me');
   });
-}
+});
