@@ -1,18 +1,10 @@
-function concat(a, b){
+export function concat(a, b){
   return `${a}${b}`;
 };
 
 // Static mutant
 const hi = '👋';
 
-function greet(name) {
+export function greet(name) {
   return `${hi} ${name}`
-}
-
-// Stryker disable all: Not useful for coverage analysis test
-if(typeof module === 'object') {
-  module.exports = {
-    concat,
-    greet
-  }
 }
