@@ -1,20 +1,11 @@
-function add(a, b) {
+export function add(a, b) {
   return a + b;
 }
 
-function multiply(a, b) {
+export function multiply(a, b) {
   return a * b;
 }
 
-function addOne(a) {
+export function addOne(a) {
   return ++a;
-}
-
-// Stryker disable all: Not useful for coverage analysis test
-if (typeof module === 'object') {
-  module.exports = {
-    add,
-    multiply,
-    addOne,
-  };
 }
