@@ -1,6 +1,5 @@
 import { expect } from 'chai';
 import { Stryker } from '@stryker-mutator/core';
-import { MutantStatus } from 'mutation-testing-report-schema/api';
 
 describe('Limit counter', () => {
   /**
@@ -11,7 +10,7 @@ describe('Limit counter', () => {
     (only ? it.only : it)(`should limit infinite loops in the ${runner}-runner`, async () => {
       const stryker = new Stryker({ testRunner: runner });
       const results = await stryker.runMutationTest();
-      const timeoutResults = results.filter((res) => res.status === MutantStatus.Timeout);
+      const timeoutResults = results.filter((res) => res.status === 'Timeout');
       expect(timeoutResults).lengthOf(3);
       timeoutResults.forEach((result) => expect(result.statusReason).eq('Hit limit reached (501/500)'));
     });
