@@ -1,3 +1,1 @@
-import * as matchers from 'vitest-dom/matchers';
-import { expect } from 'vitest';
-expect.extend(matchers);
+console.log('Empty, but here to see if stryker can deal with it');
