@@ -4,7 +4,6 @@ import {svelte} from '@sveltejs/vite-plugin-svelte'
 export default defineConfig({
   plugins: [svelte({hot: !process.env.VITEST})],
   test: {
-    environment: 'jsdom',
-    setupFiles: ['./test/vitest.setup.js'],
+    environment: 'jsdom'
   },
 })
