@@ -1,4 +1,5 @@
 import { PartialStrykerOptions } from '@stryker-mutator/api/core';
+import { Immutable } from '@stryker-mutator/util';
 
 import { CustomInitializer, CustomInitializerConfiguration } from './custom-initializer.js';
 
@@ -11,7 +12,7 @@ const guideUrl = 'https://stryker-mutator.io/docs/stryker-js/guides/vuejs';
 export class VueJsInitializer implements CustomInitializer {
   public readonly name = 'vue';
 
-  private readonly vitestConf: PartialStrykerOptions = {
+  private readonly vitestConf: Immutable<PartialStrykerOptions> = {
     testRunner: 'vitest',
     reporters: ['progress', 'clear-text', 'html'],
   };
