@@ -76,14 +76,14 @@ export class MutationTestReportHelper {
       case MutantRunStatus.Error:
         return this.reportOne({
           ...mutant,
-          status: MutantStatus.RuntimeError,
+          status: 'RuntimeError',
           statusReason: result.errorMessage,
           location,
         });
       case MutantRunStatus.Killed:
         return this.reportOne({
           ...mutant,
-          status: MutantStatus.Killed,
+          status: 'Killed',
           testsCompleted: result.nrOfTests,
           killedBy: result.killedBy,
           statusReason: result.failureMessage,
@@ -92,14 +92,14 @@ export class MutationTestReportHelper {
       case MutantRunStatus.Timeout:
         return this.reportOne({
           ...mutant,
-          status: MutantStatus.Timeout,
+          status: 'Timeout',
           statusReason: result.reason,
           location,
         });
       case MutantRunStatus.Survived:
         return this.reportOne({
           ...mutant,
-          status: MutantStatus.Survived,
+          status: 'Survived',
           testsCompleted: result.nrOfTests,
           location,
         });
@@ -114,7 +114,7 @@ export class MutationTestReportHelper {
   private checkStatusToResultStatus(status: Exclude<CheckStatus, CheckStatus.Passed>): MutantStatus {
     switch (status) {
       case CheckStatus.CompileError:
-        return MutantStatus.CompileError;
+        return 'CompileError';
     }
   }
 
