@@ -4,8 +4,6 @@ import emojiRegex from 'emoji-regex';
 
 const emojiRe = emojiRegex();
 
-const { MutantStatus } = schema;
-
 export function wrapInClosure(codeFragment: string): string {
   return `
     (function (window) {
@@ -38,20 +36,20 @@ export function deserialize<T>(stringified: string): T {
 
 export function getEmojiForStatus(status: schema.MutantStatus): string {
   switch (status) {
-    case MutantStatus.Killed:
+    case 'Killed':
       return '✅';
-    case MutantStatus.NoCoverage:
+    case 'NoCoverage':
       return '🙈';
-    case MutantStatus.Ignored:
+    case 'Ignored':
       return '🤥';
-    case MutantStatus.Survived:
+    case 'Survived':
       return '👽';
-    case MutantStatus.Timeout:
+    case 'Timeout':
       return '⏰';
-    case MutantStatus.Pending:
+    case 'Pending':
       return '⌛';
-    case MutantStatus.RuntimeError:
-    case MutantStatus.CompileError:
+    case 'RuntimeError':
+    case 'CompileError':
       return '💥';
   }
 }
