@@ -242,7 +242,7 @@ describe(ConfigReader.name, () => {
 
       // Assert
       expect(error).instanceOf(ConfigError);
-      expect(((error as ConfigError).innerError as Error).message).eq('Unexpected token n in JSON at position 2');
+      expect(((error as ConfigError).innerError as Error).message).matches(/JSON at position 2/);
     });
 
     it('should the final configuration to debug', async () => {
