@@ -3,7 +3,6 @@ import os from 'os';
 import { expect } from 'chai';
 import sinon from 'sinon';
 import chalk from 'chalk';
-import { MutantStatus } from '@stryker-mutator/api/core';
 import { factory } from '@stryker-mutator/test-helpers';
 
 import { DotsReporter } from '../../../src/reporters/dots-reporter.js';
@@ -30,7 +29,7 @@ describe(DotsReporter.name, () => {
     });
 
     it('should log "S" when status is Survived', () => {
-      sut.onMutantTested(factory.mutantResult({ status: MutantStatus.Survived }));
+      sut.onMutantTested(factory.mutantResult({ status: 'Survived' }));
       expect(process.stdout.write).to.have.been.calledWith(chalk.bold.red('S'));
     });
   });
