@@ -5,14 +5,14 @@ type ImmutablePrimitive = Primitive | ((...args: any[]) => any);
 export type Immutable<T> = T extends ImmutablePrimitive
   ? T
   : T extends Array<infer U>
-  ? ImmutableArray<U>
-  : T extends Map<infer K, infer V>
-  ? ImmutableMap<K, V>
-  : T extends Set<infer M>
-  ? ImmutableSet<M>
-  : T extends RegExp
-  ? Readonly<RegExp>
-  : ImmutableObject<T>;
+    ? ImmutableArray<U>
+    : T extends Map<infer K, infer V>
+      ? ImmutableMap<K, V>
+      : T extends Set<infer M>
+        ? ImmutableSet<M>
+        : T extends RegExp
+          ? Readonly<RegExp>
+          : ImmutableObject<T>;
 
 export type ImmutableArray<T> = ReadonlyArray<Immutable<T>>;
 export type ImmutableMap<K, V> = ReadonlyMap<Immutable<K>, Immutable<V>>;
