@@ -15,8 +15,7 @@ import {
 import { escapeRegExp, notEmpty } from '@stryker-mutator/util';
 
 import { vitestWrapper, Vitest } from './vitest-wrapper.js';
-
-import { convertTestToTestResult, fromTestId, collectTestsFromSuite, addToInlineDeps } from './vitest-helpers.js';
+import { convertTestToTestResult, fromTestId, collectTestsFromSuite, addToInlineDeps, normalizeCoverage } from './vitest-helpers.js';
 import { FileCommunicator } from './file-communicator.js';
 import { VitestRunnerOptionsWithStrykerOptions } from './vitest-runner-options-with-stryker-options.js';
 
@@ -45,8 +44,17 @@ export class VitestTestRunner implements TestRunner {
     this.setEnv();
     this.ctx = await vitestWrapper.createVitest('test', {
       config: this.options.vitest?.configFile,
+      // @ts-expect-error threads got renamed to "pool: threads" in vitest 1.0.0
       threads: true,
+      pool: 'threads',
       coverage: { enabled: false },
+      poolOptions: {
+        // Since vitest 1.0.0
+        threads: {
+          maxThreads: 1,
+          minThreads: 1,
+        },
+      },
       singleThread: false,
       maxConcurrency: 1,
       watch: false,
@@ -54,12 +62,6 @@ export class VitestTestRunner implements TestRunner {
       bail: this.options.disableBail ? 0 : 1,
       onConsoleLog: () => false,
     });
-
-    if (this.ctx.config.browser.enabled) {
-      throw new Error(
-        'Browser mode is currently not supported by the `@stryker-mutator/vitest-runner`. Please disable `browser.enabled` in your `vitest.config.js`.',
-      );
-    }
 
     // The vitest setup file needs to be inlined
     // See https://github.com/vitest-dev/vitest/issues/3403#issuecomment-1554057966
@@ -101,7 +103,7 @@ export class VitestTestRunner implements TestRunner {
     if (testIds.length > 0) {
       const regexTestNameFilter = testIds
         .map(fromTestId)
-        .map(({ name }) => escapeRegExp(name))
+        .map(({ test: name }) => escapeRegExp(name))
         .join('|');
       const regex = new RegExp(regexTestNameFilter);
       const testFiles = testIds.map(fromTestId).map(({ file }) => file);
@@ -174,7 +176,8 @@ export class VitestTestRunner implements TestRunner {
       ...new Map(this.ctx!.state.getFiles().map((file) => [`${file.projectName}-${file.name}`, file] as const)).entries(),
     ]
       .map(([, file]) => (file.meta as { mutantCoverage?: MutantCoverage }).mutantCoverage)
-      .filter(notEmpty);
+      .filter(notEmpty)
+      .map(normalizeCoverage);
 
     if (coverages.length > 1) {
       return coverages.reduce((acc, projectCoverage) => {
