@@ -3,6 +3,6 @@ import { beforeEach } from "vitest";
 import { add } from "./src/math.js";
 
 beforeEach((a) => {
-  console.log(`Detected: ${a.meta.file.name}`);
+  console.log(`Detected: ${a.task.file.name}`);
   globalThis.add = add;
 });
